+++ lib/ucx_ucc/permissions/permission.ex
@@ -15,5 +15,6 @@ defmodule UcxUcc.Permissions.Permission do
     permission
     |> cast(attrs, [:name])
     |> validate_required([:name])
+    |> unique_constraint(:name)
   end
 end