+++ priv/repo/seeds_prod.exs
@@ -17,65 +17,6 @@ Repo.delete_all Role
 Repo.delete_all Account
 Repo.delete_all User
 
-# TODO: Move this to the Permissions module so its all in one place.
-default_permissions = [
-    %{name: "access-permissions",            roles: ["admin"] },
-    %{name: "add-oauth-service",             roles: ["admin"] },
-    %{name: "add-user-to-joined-room",       roles: ["admin", "owner", "moderator"] },
-    %{name: "add-user-to-any-c-room",        roles: ["admin"] },
-    %{name: "add-user-to-any-p-room",        roles: [] },
-    %{name: "archive-room",                  roles: ["admin", "owner"] },
-    %{name: "assign-admin-role",             roles: ["admin"] },
-    %{name: "ban-user",                      roles: ["admin", "owner", "moderator"] },
-    %{name: "bulk-create-c",                 roles: ["admin"] },
-    %{name: "bulk-register-user",            roles: ["admin"] },
-    %{name: "create-c",                      roles: ["admin", "user", "bot"] },
-    %{name: "create-d",                      roles: ["admin", "user", "bot"] },
-    %{name: "create-p",                      roles: ["admin", "user", "bot"] },
-    %{name: "create-user",                   roles: ["admin"] },
-    %{name: "clean-channel-history",         roles: ["admin"] },
-    %{name: "delete-c",                      roles: ["admin"] },
-    %{name: "delete-d",                      roles: ["admin"] },
-    %{name: "delete-message",                roles: ["admin", "owner", "moderator"] },
-    %{name: "delete-p",                      roles: ["admin"] },
-    %{name: "delete-user",                   roles: ["admin"] },
-    %{name: "edit-message",                  roles: ["admin", "owner", "moderator"] },
-    %{name: "edit-other-user-active-status", roles: ["admin"] },
-    %{name: "edit-other-user-info",          roles: ["admin"] },
-    %{name: "edit-other-user-password",      roles: ["admin"] },
-    %{name: "edit-privileged-setting",       roles: ["admin"] },
-    %{name: "edit-room",                     roles: ["admin", "owner", "moderator"] },
-    %{name: "manage-assets",                 roles: ["admin"] },
-    %{name: "manage-emoji",                  roles: ["admin"] },
-    %{name: "manage-integrations",           roles: ["admin"] },
-    %{name: "manage-own-integrations",       roles: ["admin", "bot"] },
-    %{name: "manage-oauth-apps",             roles: ["admin"] },
-    %{name: "mention-all",                   roles: ["admin", "owner", "moderator", "user"] },
-    %{name: "mute-user",                     roles: ["admin", "owner", "moderator"] },
-    %{name: "pin-message",                   roles: ["admin", "owner", "moderator"] },
-    %{name: "remove-user",                   roles: ["admin", "owner", "moderator"] },
-    %{name: "run-import",                    roles: ["admin"] },
-    %{name: "run-migration",                 roles: ["admin"] },
-    %{name: "set-moderator",                 roles: ["admin", "owner"] },
-    %{name: "set-owner",                     roles: ["admin", "owner"] },
-    %{name: "unarchive-room",                roles: ["admin"] },
-    %{name: "view-c-room",                   roles: ["admin", "user", "bot"] },
-    %{name: "view-d-room",                   roles: ["admin", "user", "bot"] },
-    %{name: "view-full-other-user-info",     roles: ["admin"] },
-    %{name: "view-history",                  roles: ["admin", "user"] },
-    %{name: "view-joined-room",              roles: ["guest", "bot"] },
-    %{name: "view-join-code",                roles: ["admin"] },
-    %{name: "view-logs",                     roles: ["admin"] },
-    %{name: "view-other-user-channels",      roles: ["admin"] },
-    %{name: "view-p-room",                   roles: ["admin", "user"] },
-    %{name: "view-privileged-setting",       roles: ["admin"] },
-    %{name: "view-room-administration",      roles: ["admin"] },
-    %{name: "view-message-administration",   roles: ["admin"] },
-    %{name: "view-statistics",               roles: ["admin"] },
-    %{name: "view-user-administration",      roles: ["admin"] },
-    %{name: "preview-c-room",                roles: ["admin", "user"] }
-  ]
-
 IO.puts "Creating Roles"
 
 roles =
@@ -123,8 +64,11 @@ end
 
 IO.puts "Creating Permissions"
 # build the permissions
+Repo.delete_all UcxUcc.Permissions.Permission
+
 roles_list = roles
-default_permissions
+
+UcxUcc.Permissions.default_permissions()
 |> Enum.each(fn %{name: name, roles: roles} ->
   {:ok, permission} = Permissions.create_permission(%{name: name})
   roles